tests/container/requirements.txt
google-cloud-logging
gunicorn
flask
